[Refactor] Restructure the /auth/user/ apis.
