+++ src/main/java/top/horizonask/hoawiki/authentication/controller/info/PermissionController.java
@@ -1,4 +1,4 @@
-package top.horizonask.hoawiki.authentication.controller.permission;
+package top.horizonask.hoawiki.authentication.controller.info;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -7,6 +7,7 @@ import top.horizonask.hoawiki.common.ApiResponse;
 import top.horizonask.hoawiki.authentication.entity.Permission;
 import top.horizonask.hoawiki.authentication.mapper.PermissionMapper;
 
+import javax.management.relation.RoleNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,6 +19,7 @@ import java.util.List;
 
 @Slf4j
 @RestController
+@RequestMapping(value ="/auth/user/info")
 public class PermissionController {
     public PermissionMapper permissionMapper;
 
@@ -27,10 +29,13 @@ public class PermissionController {
 
     @PreAuthorize(value="hasRole('Admin') or hasPermission('/permission/role','admin:rolesPermission')")
     @RequestMapping("/permission/role/{roleId}")
-    public ApiResponse permissionId(@PathVariable Long roleId) {
+    public ApiResponse permissionId(@PathVariable Long roleId) throws RoleNotFoundException {
         List<Long> roleIds = new ArrayList<>();
         roleIds.add(roleId);
         List<Permission> permission = permissionMapper.findByRoleId(roleIds);
+        if(permission == null){
+            throw new RoleNotFoundException();
+        }
         ApiResponse res = ApiResponse.ok().message("Get user information");
         for (Permission
                 onePermission :
